work on app style
