+++ public/index.html
@@ -13,8 +13,6 @@
 
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
-    <link rel="stylesheet" href="./style/form.css">
-   
     <title>React App</title>
   </head>
   <body>
