+++ src/App.js
@@ -1,13 +1,13 @@
-import logo from './logo.svg';
 import './App.css';
+import FormComponent from './FormComponent/FormComponent';
 
-import FormComponent from './components/FormComponent/FormComponent';
+
 
 function App() {
   return (
     <div className="App">
-
-      <FormComponent/>
+      <FormComponent />
+    
     </div>
   );
 }
