+++ src/FormComponent/FormComponents.css
@@ -8,22 +8,16 @@
     line-height: 1.3;
 }
 
-
-body{
-    background-image: url(../assets/bg-fixed-3.jpg);
+body {
+    background-image: url(/public/bg-fixed-3.jpg);
     color: white;
 }
 
-
-#campaign{
-
+#campaign {
     align-items: center;
-
 }
 
-
-
-.container{
+.container {
     position: relative;
     width: 800px;
     height: auto;
@@ -35,10 +29,17 @@ body{
     box-shadow: 0px 0px 32px 0px #32485e, -14px -14px 20px #191c21;
 }
 
-.label-title{
+.label-title {
     display: block;
     margin-top: 20px;
     font-size: 25px;
+}
+
+#campaign-description {
+    margin-top: 10px;
+    width: 350px;
+    height: 200px;
+    border-radius: 10px;
 }
 
 
@@ -50,32 +51,48 @@ input[type="text"] {
     margin-top: 10px;
     font-size: 24px;
     border-radius: 50px;
-
 }
 
-input[type="data"]{
+input[type="date"] {
+    display: block;
     background: #ecf0f3;
-    padding: 30px;
+    padding: 20px;
     padding-left: 20px;
     height: 50px;
     margin-top: 10px;
-    font-size: 24px;
+    font-size: 20px;
     border-radius: 50px;
-
 }
 
-.buttons{
+.buttons {
     display:block;
     margin-top: 10px;
 
-
 }
 
-.radio-btn-container{
-    width: 150px;
-    text-align: initial
-}
-
-.label-language{
+.paid-decision { 
+    padding-left: 20px;
     font-size: 20px;
+}
+
+#languages {
+    width:  300px;
+    height: 30px;
+}
+
+.language-option {
+    font-size: 17px;
+    margin-top: 10px;
+}
+
+#submit-btn {
+    display: block;
+    background: #778e9c;
+    margin-top: 30px;
+    height: 60px;
+    width: 200px;
+    border-radius: 50px;   
+    color: white; 
+    font-size: 22px;
+    cursor: pointer;
 }
